requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==17.6.0
kombu==5.2.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.6
